feat: This commit adds code changes that are required for creating ConfigMaps in remote cluster
This commit address a number of things. Firstly it adds new Annotation named cross-cluster-type
in Kafka CR and new label named submariner-cluster-id in KNP CR. This is required for Cross Cluster
communication when Brokers and controllers are deployed in Stretch Cluster Mode. In addition to this
It fixes an issue with Creating ServiceAccount in Remote Clusters. The major outcome of this commit is
it modifies the advertised.listener and Controller.quorum.voters properties that Kafka uses for interbroker
communication (Replication) and Leader election respectively, to have Submariner exported Headless Service DNS entry

Signed-off-by: Aswin A <aswin6303@gmail.com>